client/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "sdp-client"
version = "0.2.0"
description = "Secure Data Processing client CLI"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "requests>=2.31",
  "cryptography>=42",
  "pyyaml>=6.0",
]

[project.scripts]
sdp-client = "sdp_client.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}
include-package-data = true

[tool.setuptools.packages.find]
where = ["src"]
include = ["sdp_client*"]
